Refactor: Move API configs to own file
These configs are shared between SDK and multitenant SDK.
